refactor: move selection rendering to dedicated UILayer
