+++ src/client/graphics/layers/UnitLayer.ts
@@ -3,7 +3,11 @@ import { Theme } from "../../../core/configuration/Config";
 import { Unit, UnitType, Player } from "../../../core/game/Game";
 import { Layer } from "./Layer";
 import { EventBus } from "../../../core/EventBus";
-import { AlternateViewEvent, MouseUpEvent } from "../../InputHandler";
+import {
+  AlternateViewEvent,
+  MouseUpEvent,
+  UnitSelectionEvent,
+} from "../../InputHandler";
 import { ClientID } from "../../../core/Schemas";
 import { GameView, PlayerView, UnitView } from "../../../core/game/GameView";
 import {
@@ -12,7 +16,6 @@ import {
   TileRef,
 } from "../../../core/game/GameMap";
 import { GameUpdateType } from "../../../core/game/GameUpdates";
-import { SelectedUnits, UnitSelectionEvent } from "../SelectedUnits";
 import { TransformHandler } from "../TransformHandler";
 
 enum Relationship {
@@ -35,9 +38,10 @@ export class UnitLayer implements Layer {
 
   private oldShellTile = new Map<UnitView, TileRef>();
 
-  private selectedUnits: SelectedUnits;
-  private selectionAnimTime = 0;
   private transformHandler: TransformHandler;
+
+  // Selected unit property as suggested in the review comment
+  private selectedUnit: UnitView | null = null;
 
   // Configuration for unit selection
   private readonly WARSHIP_SELECTION_RADIUS = 3; // Radius in game cells for warship selection hit zone
@@ -49,7 +53,6 @@ export class UnitLayer implements Layer {
     transformHandler: TransformHandler,
   ) {
     this.theme = game.config().theme();
-    this.selectedUnits = new SelectedUnits(eventBus);
     this.transformHandler = transformHandler;
   }
 
@@ -64,20 +67,12 @@ export class UnitLayer implements Layer {
     this.game.updatesSinceLastTick()?.[GameUpdateType.Unit]?.forEach((unit) => {
       this.onUnitEvent(this.game.unit(unit.id));
     });
-
-    // Update the selection animation time
-    this.selectionAnimTime = (this.selectionAnimTime + 1) % 60;
-
-    // If there's a selected warship, redraw to update the selection box animation
-    if (this.selectedUnits.hasSelectedUnitOfType(UnitType.Warship)) {
-      this.redrawSelectedUnit(this.selectedUnits.getSelectedUnit());
-    }
   }
 
   init() {
     this.eventBus.on(AlternateViewEvent, (e) => this.onAlternativeViewEvent(e));
     this.eventBus.on(MouseUpEvent, (e) => this.onMouseUp(e));
-    this.eventBus.on(UnitSelectionEvent, (e) => this.onUnitSelection(e));
+    this.eventBus.on(UnitSelectionEvent, (e) => this.onUnitSelectionChange(e));
     this.redraw();
   }
 
@@ -116,48 +111,29 @@ export class UnitLayer implements Layer {
     const nearbyWarships = this.findWarshipsNearCell(cell);
 
     if (nearbyWarships.length > 0) {
-      // Select/deselect the closest warship
-      this.selectedUnits.toggleUnitSelection(nearbyWarships[0]);
-    } else if (this.selectedUnits.getSelectedUnit()) {
+      // Toggle selection of the closest warship
+      const clickedUnit = nearbyWarships[0];
+      if (this.selectedUnit === clickedUnit) {
+        // Deselect if already selected
+        this.eventBus.emit(new UnitSelectionEvent(clickedUnit, false));
+      } else {
+        // Select the new unit
+        this.eventBus.emit(new UnitSelectionEvent(clickedUnit, true));
+      }
+    } else if (this.selectedUnit) {
       // If clicked elsewhere and there's a selection, deselect it
-      this.selectedUnits.deselectCurrentUnit();
+      this.eventBus.emit(new UnitSelectionEvent(this.selectedUnit, false));
     }
   }
 
   /**
-   * Clear the selection box at a specific position
+   * Handle unit selection changes
    */
-  private clearSelectionBox(x: number, y: number, size: number) {
-    for (let px = x - size; px <= x + size; px++) {
-      for (let py = y - size; py <= y + size; py++) {
-        if (
-          px === x - size ||
-          px === x + size ||
-          py === y - size ||
-          py === y + size
-        ) {
-          this.clearCell(px, py);
-        }
-      }
-    }
-  }
-
-  private onUnitSelection(event: UnitSelectionEvent) {
-    if (event.unit && event.unit.type() === UnitType.Warship) {
-      if (event.isSelected) {
-        // Highlight the selected warship
-        this.redrawSelectedUnit(event.unit);
-      } else {
-        // Remove the highlight
-        this.onUnitEvent(event.unit);
-
-        // Also clear any lingering selection box
-        if (this.lastSelectionBoxCenter) {
-          const { x, y, size } = this.lastSelectionBoxCenter;
-          this.clearSelectionBox(x, y, size);
-          this.lastSelectionBoxCenter = null;
-        }
-      }
+  private onUnitSelectionChange(event: UnitSelectionEvent) {
+    if (event.isSelected) {
+      this.selectedUnit = event.unit;
+    } else if (this.selectedUnit === event.unit) {
+      this.selectedUnit = null;
     }
   }
 
@@ -166,24 +142,8 @@ export class UnitLayer implements Layer {
    * If the selected unit is removed from the game, deselect it
    */
   private handleUnitDeactivation(unit: UnitView) {
-    if (this.selectedUnits.isSelected(unit) && !unit.isActive()) {
-      // Clear the selection box before deselecting
-      if (
-        this.lastSelectionBoxCenter &&
-        this.lastSelectionBoxCenter.unit === unit
-      ) {
-        const { x, y, size } = this.lastSelectionBoxCenter;
-        this.clearSelectionBox(x, y, size);
-        this.lastSelectionBoxCenter = null;
-      }
-
-      this.selectedUnits.deselectCurrentUnit();
-    }
-  }
-
-  private redrawSelectedUnit(unit: UnitView) {
-    if (unit && unit.type() === UnitType.Warship && unit.isActive()) {
-      this.onUnitEvent(unit);
+    if (this.selectedUnit === unit && !unit.isActive()) {
+      this.eventBus.emit(new UnitSelectionEvent(unit, false));
     }
   }
 
@@ -308,98 +268,6 @@ export class UnitLayer implements Layer {
         255,
       );
     }
-
-    // If this is a selected warship, draw the selection box
-    if (this.selectedUnits.isSelected(unit)) {
-      this.drawSelectionBox(unit);
-    }
-  }
-
-  // Keep track of previous selection box positions for cleanup
-  private lastSelectionBoxCenter: {
-    unit: UnitView;
-    x: number;
-    y: number;
-    size: number;
-  } | null = null;
-
-  // Visual settings for selection
-  private readonly SELECTION_BOX_SIZE = 6; // Size of the selection box (should be larger than the warship)
-
-  /**
-   * Draw a selection box around the warship
-   */
-  private drawSelectionBox(unit: UnitView) {
-    // Use the configured selection box size
-    const selectionSize = this.SELECTION_BOX_SIZE;
-
-    // Calculate pulsating effect based on animation time (25% variation in opacity)
-    const baseOpacity = 200;
-    const pulseAmount = 55;
-    const opacity =
-      baseOpacity + Math.sin(this.selectionAnimTime * 0.1) * pulseAmount;
-
-    // Get the warship's owner color for the box
-    const ownerColor = this.theme.territoryColor(unit.owner().info());
-
-    // Create a brighter version of the owner color for the selection
-    const selectionColor = ownerColor.lighten(0.2);
-
-    // Get current center position
-    const center = unit.tile();
-    const centerX = this.game.x(center);
-    const centerY = this.game.y(center);
-
-    // Clear previous selection box if it exists and is different from current position
-    if (
-      this.lastSelectionBoxCenter &&
-      (this.lastSelectionBoxCenter.x !== centerX ||
-        this.lastSelectionBoxCenter.y !== centerY ||
-        this.lastSelectionBoxCenter.unit !== unit)
-    ) {
-      const lastSize = this.lastSelectionBoxCenter.size;
-      const lastX = this.lastSelectionBoxCenter.x;
-      const lastY = this.lastSelectionBoxCenter.y;
-
-      // Clear the previous selection box
-      this.clearSelectionBox(lastX, lastY, lastSize);
-
-      // We don't need to redraw the territory since the unit layer sits on top of the territory layer
-      // and clearing just the selection box pixels won't affect the territory underneath
-    }
-
-    // Draw the selection box
-    for (let x = centerX - selectionSize; x <= centerX + selectionSize; x++) {
-      for (let y = centerY - selectionSize; y <= centerY + selectionSize; y++) {
-        // Only draw if it's on the border (not inside or outside the box)
-        if (
-          x === centerX - selectionSize ||
-          x === centerX + selectionSize ||
-          y === centerY - selectionSize ||
-          y === centerY + selectionSize
-        ) {
-          // Create a dashed effect by only drawing some pixels
-          const dashPattern = (x + y) % 2 === 0;
-          if (dashPattern) {
-            this.paintCell(
-              x,
-              y,
-              this.relationship(unit),
-              selectionColor,
-              opacity,
-            );
-          }
-        }
-      }
-    }
-
-    // Store current selection box position for next cleanup
-    this.lastSelectionBoxCenter = {
-      unit,
-      x: centerX,
-      y: centerY,
-      size: selectionSize,
-    };
   }
 
   private handleShellEvent(unit: UnitView) {
